convert away from hydrated react, and then use it
