+++ apps/prairielearn/src/pages/studentInstanceQuestion/studentInstanceQuestion.html.ts
@@ -196,7 +196,7 @@ export function StudentInstanceQuestion({
             variant: resLocals.variant,
             authz_result: resLocals.authz_result,
             csrfToken: resLocals.__csrf_token,
-            urlPrefix: resLocals.urlPrefix,
+            courseInstanceId: resLocals.course_instance.id,
             allowGradeLeftMs: resLocals.allowGradeLeftMs,
           })}
           ${QuestionNavSideGroup({
